continued refactoring into separate files
