+++ ubyssey/static_src/src/stove/assignment_manager/StoryTypeOptions.json
@@ -37,7 +37,7 @@
   {
     "value": "commentary", 
     "label": "Commentary", 
-    "description": "[Description TK]"
+    "description": "A commentary is driven by an author's perspective on the sport they cover as a beat reporter. They take a position on an event or topic relevant to the sport, and suggest a course of action, either in future or past-tense."
   },
   {
     "value": "essay",
